+++ src/components/ChannelList/__tests__/ChannelList.test.js
@@ -1,3 +1,4 @@
+/* eslint-disable sonarjs/no-duplicate-string */
 import React from 'react';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { getNodeText } from '@testing-library/dom';
@@ -20,6 +21,8 @@ import {
   dispatchMessageNewEvent,
   dispatchChannelDeletedEvent,
   dispatchChannelUpdatedEvent,
+  dispatchChannelHiddenEvent,
+  dispatchChannelVisibleEvent,
   dispatchChannelTruncatedEvent,
   dispatchNotificationAddedToChannelEvent,
   dispatchNotificationMessageNewEvent,
@@ -777,6 +780,141 @@ describe('ChannelList', () => {
 
         await waitFor(() => {
           expect(setActiveChannel).toHaveBeenCalledTimes(1);
+        });
+      });
+    });
+
+    describe('channel.hidden', () => {
+      const channelListProps = {
+        filters: {},
+        Preview: ChannelPreviewComponent,
+        List: ChannelListComponent,
+      };
+
+      // eslint-disable-next-line sonarjs/no-identical-functions
+      beforeEach(() => {
+        useMockedApis(chatClientUthred, [
+          queryChannelsApi([testChannel1, testChannel2]),
+        ]);
+      });
+
+      it('should remove channel from list, by default', async () => {
+        const { getByRole, getByTestId } = render(
+          <Chat client={chatClientUthred}>
+            <ChannelList {...channelListProps} />
+          </Chat>,
+        );
+
+        // Wait for list of channels to load in DOM.
+        await waitFor(() => {
+          expect(getByRole('list')).toBeInTheDocument();
+        });
+
+        const nodeToBeRemoved = getByTestId(testChannel2.channel.id);
+        act(() =>
+          dispatchChannelHiddenEvent(chatClientUthred, testChannel2.channel),
+        );
+
+        await waitFor(() => {
+          expect(nodeToBeRemoved).not.toBeInTheDocument();
+        });
+      });
+
+      it('should unset activeChannel if it was hidden', async () => {
+        const setActiveChannel = jest.fn();
+        const { getByRole } = render(
+          <ChatContext.Provider
+            value={{ client: chatClientUthred, setActiveChannel }}
+          >
+            <ChannelList
+              {...channelListProps}
+              channel={{ cid: testChannel1.channel.cid }}
+              setActiveChannel={setActiveChannel}
+            />
+          </ChatContext.Provider>,
+        );
+
+        // Wait for list of channels to load in DOM.
+        await waitFor(() => {
+          expect(getByRole('list')).toBeInTheDocument();
+        });
+
+        act(() =>
+          dispatchChannelHiddenEvent(chatClientUthred, testChannel1.channel),
+        );
+
+        await waitFor(() => {
+          expect(setActiveChannel).toHaveBeenCalledTimes(1);
+        });
+      });
+    });
+
+    describe('channel.visible', () => {
+      const channelListProps = {
+        filters: {},
+        Preview: ChannelPreviewComponent,
+        List: ChannelListComponent,
+        options: { state: true, watch: true, presence: true },
+      };
+
+      beforeEach(async () => {
+        chatClientUthred = await getTestClientWithUser({ id: 'jaap' });
+        useMockedApis(chatClientUthred, [
+          queryChannelsApi([testChannel1, testChannel2]),
+          getOrCreateChannelApi(testChannel3),
+        ]);
+      });
+
+      it('should move channel to top of the list by default', async () => {
+        const { getByRole, getByTestId, getAllByRole } = render(
+          <Chat client={chatClientUthred}>
+            <ChannelList {...channelListProps} />
+          </Chat>,
+        );
+
+        // Wait for list of channels to load in DOM.
+        await waitFor(() => {
+          expect(getByRole('list')).toBeInTheDocument();
+        });
+
+        act(() =>
+          dispatchChannelVisibleEvent(chatClientUthred, testChannel3.channel),
+        );
+
+        await waitFor(() => {
+          expect(getByTestId(testChannel3.channel.id)).toBeInTheDocument();
+        });
+
+        const items = getAllByRole('listitem');
+
+        // Get the closes listitem to the channel that received new message.
+        const channelPreview = getByTestId(testChannel3.channel.id);
+        expect(channelPreview.isEqualNode(items[0])).toBe(true);
+      });
+
+      it('should call `onChannelVisible` function prop, if provided', async () => {
+        const onChannelVisible = jest.fn();
+        const { getByRole } = render(
+          <Chat client={chatClientUthred}>
+            <ChannelList
+              {...channelListProps}
+              onChannelVisible={onChannelVisible}
+            />
+          </Chat>,
+        );
+
+        // Wait for list of channels to load in DOM.
+        await waitFor(() => {
+          expect(getByRole('list')).toBeInTheDocument();
+        });
+
+        // eslint-disable-next-line sonarjs/no-identical-functions
+        act(() =>
+          dispatchChannelVisibleEvent(chatClientUthred, testChannel3.channel),
+        );
+
+        await waitFor(() => {
+          expect(onChannelVisible).toHaveBeenCalledTimes(1);
         });
       });
     });