+++ src/mock-builders/event/connectionChanged.js
@@ -1,0 +1,6 @@
+export default (client, online) => {
+  client.dispatchEvent({
+    type: 'connection.changed',
+    online,
+  });
+};